+++ .php-cs-fixer.dist.php
@@ -45,20 +45,20 @@ return (new PhpCsFixer\Config())
         (new PhpCsFixer\Finder())
             ->in(__DIR__ . '/src')
             ->exclude([
-                'bb-data',
-                'bb-library',
-                'bb-locale',
-                'bb-themes',
+                'data',
+                'library',
+                'locale',
+                'themes',
                 'vendor',
                 'install',
-                'bb-modules/Wysiwyg',
-                'bb-modules/Servicecentovacast',
-                'bb-modules/Servicesolusvm',
-                'bb-modules/Spamchecker'
+                'modules/Wysiwyg',
+                'modules/Servicecentovacast',
+                'modules/Servicesolusvm',
+                'modules/Spamchecker'
             ])
             ->notPath(
                 'rb.php',
-                'bb-config-sample.php'
+                'config-sample.php'
             )
     )
 ;