+++ tests/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 define('APPLICATION_ENV', 'testing');
 define('BB_PATH_TESTS', dirname(__FILE__));
-require_once dirname(__FILE__) . '/../src/bb-load.php';
+require_once dirname(__FILE__) . '/../src/load.php';
 $config = include dirname(__FILE__) . '/../src/bb-config.php';
 
 require_once dirname(__FILE__) . '/../src/vendor/autoload.php';
@@ -17,7 +17,7 @@ define('BB_DB_TYPE', $config['db']['type']);
 // Add test libraries
 set_include_path(implode(PATH_SEPARATOR, array(
     get_include_path(),
-    BB_PATH_TESTS.'/bb-library',
+    BB_PATH_TESTS.'/library',
     BB_PATH_TESTS.'/includes',
     BB_PATH_TESTS.'/includes/Vps',
 )));
@@ -32,5 +32,5 @@ require_once 'BBModTestCase.php';
 require_once BB_PATH_TESTS.'/includes/Payment/Adapter/Dummy.php';
 require_once 'FakeTemplateWrapper.php';
 /**/
-$di = include BB_PATH_ROOT . '/bb-di.php';
+$di = include BB_PATH_ROOT . '/di.php';
 $di['translate']();